Added an initial version of a transcoding (updated)

+++ vk_video_encoder/libs/VkVideoEncoder/VkVideoGopStructure.cpp
@@ -16,7 +16,7 @@
 
 #include "VkVideoGopStructure.h"
 
-VkVideoGopStructure::VkVideoGopStructure(uint8_t gopFrameCount,
+VkVideoGopStructure::VkVideoGopStructure(uint16_t gopFrameCount,
                                          int32_t idrPeriod,
                                          uint8_t consecutiveBFrameCount,
                                          uint8_t temporalLayerCount,
@@ -38,7 +38,7 @@ VkVideoGopStructure::VkVideoGopStructure(uint8_t gopFrameCount,
 bool VkVideoGopStructure::Init(uint64_t maxNumFrames)
 {
     m_gopFrameCycle = (uint8_t)(m_consecutiveBFrameCount + 1);
-    m_gopFrameCount = (uint8_t)std::min<uint64_t>(m_gopFrameCount, maxNumFrames);
+    m_gopFrameCount = (uint16_t)std::min<uint64_t>(m_gopFrameCount, maxNumFrames);
     if (m_idrPeriod > 0) {
         m_idrPeriod = (uint32_t)std::min<uint64_t>(m_idrPeriod, maxNumFrames);
     }
